refactor(core): share symbol table writing between RISC-V and Pep/10
